test: split sonar coverage wave into domain test groups
