+++ config/pytest-groups/light.txt
@@ -205,3 +205,9 @@ tests/test_medium_gap_modules_coverage.py
 tests/test_orchestrator_compat_module.py
 tests/test_strategy_and_routes_hotfix.py
 tests/test_orchestrator_flows_hotfix.py
+tests/test_models_remote_light_coverage.py
+tests/test_service_monitor_light_coverage.py
+tests/test_agents_skill_manager_paths.py
+tests/test_runtime_registry_light_coverage.py
+tests/test_governance_light_coverage.py
+tests/test_academy_conversion_light_coverage.py
